Added 2 Random Values to Pathmaker Start & Began Scene Reloading Script
Now each new Pathmaker randomizes the probability that it will generate normal tiles (versus the fancy ones I made in Maya), as well as the length of the halls it will make.
So far the scene reloading script reloads the scene, but won't allow the pathmaker to start up again.